Book api consumer

+++ DataAccessLayer/Entities/Publisher.cs
@@ -8,6 +8,7 @@ namespace Contract.Entities
 {
     public class Publisher
     {
+        // diable auto generated 
         public int Id { get; set; }
         [Required]
         [StringLength(25, MinimumLength = 2)]
@@ -22,7 +23,7 @@ namespace Contract.Entities
         public DateTime DateOfBirth { get; set; }
         [Required]
         [Range(2000,5000,ErrorMessage="it must be between 2000 and 5000 $")]
-        public int Salery { get; set; }
+        public float Salery { get; set; }
         //Navigation Properites
 
         public List<Book> Books { get; set; }
